The program starts recording when it detects a face or a body through the CasacadeClassifier and ends the same when the condition becomes false. When at a closer proximity to a body, the program issues alert to the user.
Runs perfectly on Conda environment, Python 3.9 given OpenCV module has been installed.
A beginner's code - The efficiency of the code depends on the system as well as the camera being used since the maximum frames per seconds matters.
CascadeClassifier has been used to identify faces and bodies
Videos are saved in mp4 format and the specific folder where the videos are to be saved can be specified in VScode.